refactor(telegram): extract /status feature into its own module
CI's `bun run check` was red on `scripts/check-file-lines.mjs`:
`handlers.py` hit 666 lines after the /status command commit pushed
it past the project's 500-line ceiling. Pulled the entire /status
feature (formatters, render helper, async handler, and its copy
constants) into `app/integrations/telegram/status.py` so each module
holds one concern.

The shape:

* `status.py` now owns `_VERBOSE_LABELS` (used by both /status's
  `_resolve_verbose` and `handle_verbose_command` in handlers.py).
  Defining it here avoids the `handlers → status → handlers` import
  cycle that would form if it lived in handlers; the verbose command
  imports it back.
* `handlers.py` shrinks to 489 lines and only keeps the message-flow
  handlers (`/start`, plain text, `/new`, `/stop`, `/model`,
  `/verbose`).
* `bot.py` and `test_telegram_channel.py` import `handle_status_command`
  directly from `.status`; the three TestHandleStatusCommand patches
  now target `app.integrations.telegram.status.*` so they match where
  the calls actually resolve.

793 backend tests green (full suite); ruff + format clean; file-line
and nesting budgets clean; frontend bun-check green.
