+++ backend/app/integrations/telegram/handlers.py
@@ -28,14 +28,12 @@ from __future__ import annotations
 import logging
 import re
 import uuid
-from collections.abc import Callable
 from dataclasses import dataclass
-from datetime import UTC, datetime
+from datetime import datetime
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.config import settings
-from app.core.providers.catalog import default_model, find
+from app.core.providers.catalog import default_model
 from app.core.providers.model_id import InvalidModelId, parse_model_id
 from app.crud.channel import (
     get_or_create_telegram_conversation_full,
@@ -44,7 +42,6 @@ from app.crud.channel import (
     update_conversation_model,
     update_conversation_verbose_level,
 )
-from app.crud.conversation import ConversationStatus, get_conversation_status
 
 # Loose match for the link-code shape (8 chars from the look-alike-free
 # alphabet defined in app.crud.channel). Used to distinguish "user pasted
@@ -101,89 +98,13 @@ _VERBOSE_NOT_BOUND_MESSAGE = "Connect your account first before changing the ver
 _VERBOSE_OK_MESSAGE = "Verbose level set to {level} ({label})."
 _VERBOSE_FAIL_MESSAGE = "Couldn't update verbose level — please try again."
 
-# Human-readable labels used in the /verbose reply.
-_VERBOSE_LABELS: dict[int, str] = {
-    0: "quiet",
-    1: "normal",
-    2: "detailed",
-}
-
-# /status handler copy
-_STATUS_NOT_BOUND_MESSAGE = "Connect your account first before asking for status."
-_STATUS_NO_CONVERSATION_MESSAGE = (
-    "📊 Pawrrtal gateway\n\n"
-    "⏱  Bot up: {uptime} (this worker)\n\n"
-    "💬 No conversation yet — send a message to start one."
+# The ``/status`` feature (formatters + handler + its copy constants)
+# lives in :mod:`app.integrations.telegram.status` to keep this file
+# under the file-line budget.  Re-exported below so existing callers
+# (``bot.py``, tests) keep importing from ``handlers``.
+from app.integrations.telegram.status import (  # noqa: E402
+    _VERBOSE_LABELS,
 )
-_STATUS_MESSAGE = (
-    "📊 Pawrrtal gateway\n\n"
-    "⏱  Bot up: {uptime} (this worker)\n"
-    "🤖 Model: {model_display} (<code>{model_id}</code>){model_warning}\n"
-    "🔊 Verbose: {verbose_level} ({verbose_label})\n\n"
-    "💬 This conversation\n"
-    "   • Started: {started_ago} ago\n"
-    "   • Messages: {messages} ({user_messages} yours, {assistant_messages} assistant)\n"
-    "   • Tokens: {tokens}\n"
-    "   • Status: {run_status}"
-)
-_STATUS_THREAD_LINE = "\n🧵 Topic thread: <code>{thread_id}</code>"
-_STATUS_MODEL_WARNING_SUFFIX = " ⚠️ catalog lookup failed"
-_STATUS_RUN_RUNNING = "running"
-_STATUS_RUN_IDLE = "idle"
-
-
-def _format_duration(seconds: float) -> str:
-    """Render a positive duration as ``"3d 1h"`` / ``"4h 12m"`` / ``"34s"``.
-
-    Falls back to ``"0s"`` for non-positive inputs so the formatter never
-    yields an empty string.
-    """
-    total = int(max(0.0, seconds))
-    days, rem = divmod(total, 86_400)
-    hours, rem = divmod(rem, 3_600)
-    minutes, secs = divmod(rem, 60)
-    if days:
-        return f"{days}d {hours}h"
-    if hours:
-        return f"{hours}h {minutes}m"
-    if minutes:
-        return f"{minutes}m {secs}s"
-    return f"{secs}s"
-
-
-def _format_token_count(n: int) -> str:
-    """Render a non-negative token count with thousands separators."""
-    return f"{max(0, int(n)):,}"
-
-
-def _format_model_display(model_id: str | None) -> tuple[str, str, str]:
-    """Resolve a stored model_id to a (display_name, canonical_id, warning) triple.
-
-    Falls back to the catalog default when ``model_id`` is ``None``. When the
-    ID is set but absent from the catalog, returns the raw ID with a warning
-    suffix so the status reply still renders without crashing.
-    """
-    canonical = model_id or default_model().id
-    try:
-        parsed = parse_model_id(canonical)
-        entry = find(parsed)
-    except InvalidModelId:
-        entry = None
-
-    if entry is None:
-        return canonical, canonical, _STATUS_MODEL_WARNING_SUFFIX
-    return entry.short_name, entry.id, ""
-
-
-def _resolve_verbose(level: int | None) -> tuple[int, str]:
-    """Return ``(effective_level, label)`` for the status reply."""
-    effective = level if level is not None else int(settings.telegram_verbose_default)
-    return effective, _VERBOSE_LABELS.get(effective, "unknown")
-
-
-def _now_utc() -> datetime:
-    """Indirection seam for tests that want to freeze 'now'."""
-    return datetime.now(UTC)
 
 
 @dataclass(frozen=True)
@@ -376,8 +297,6 @@ async def handle_new_command(
     if nexus_user_id is None:
         return _NEW_NOT_BOUND_MESSAGE
 
-    from datetime import datetime  # noqa: PLC0415 — already imported by callers
-
     from app.models import Conversation  # noqa: PLC0415
 
     conversation = Conversation(
@@ -564,103 +483,6 @@ async def handle_verbose_command(
     return _VERBOSE_OK_MESSAGE.format(level=level, label=_VERBOSE_LABELS[level])
 
 
-def _render_status_message(
-    *,
-    bot_uptime_seconds: float,
-    status: ConversationStatus,
-    run_active: bool,
-    thread_id: int | None,
-    now: datetime,
-) -> str:
-    """Pure formatter used by ``handle_status_command`` and its tests."""
-    model_display, model_canonical, model_warning = _format_model_display(status.model_id)
-    verbose_level, verbose_label = _resolve_verbose(status.verbose_level)
-    # ``Conversation.created_at`` is a tz-naive DateTime column in the DB —
-    # all timestamps in this app are written in UTC. Normalize so the
-    # subtraction never trips on mixed naive/aware operands.
-    started_at = status.started_at
-    if started_at.tzinfo is None:
-        started_at = started_at.replace(tzinfo=UTC)
-    started_ago_seconds = (now - started_at).total_seconds()
-
-    # Tokens come from cost_ledger, which is only populated for providers
-    # that emit ``usage`` stream events. Gemini currently doesn't, so its
-    # turns land with zero tokens despite real activity. Render an honest
-    # placeholder rather than a misleading "0 in / 0 out" in that case.
-    has_messages = status.message_count > 0
-    has_usage = status.total_input_tokens > 0 or status.total_output_tokens > 0
-    if has_messages and not has_usage:
-        tokens_line = "n/a (provider did not report usage)"
-    else:
-        tokens_line = (
-            f"{_format_token_count(status.total_input_tokens)} in / "
-            f"{_format_token_count(status.total_output_tokens)} out"
-        )
-
-    body = _STATUS_MESSAGE.format(
-        uptime=_format_duration(bot_uptime_seconds),
-        model_display=model_display,
-        model_id=model_canonical,
-        model_warning=model_warning,
-        verbose_level=verbose_level,
-        verbose_label=verbose_label,
-        started_ago=_format_duration(started_ago_seconds),
-        messages=status.message_count,
-        user_messages=status.user_message_count,
-        assistant_messages=status.assistant_message_count,
-        tokens=tokens_line,
-        run_status=_STATUS_RUN_RUNNING if run_active else _STATUS_RUN_IDLE,
-    )
-    if thread_id is not None:
-        body += _STATUS_THREAD_LINE.format(thread_id=thread_id)
-    return body
-
-
-async def handle_status_command(
-    *,
-    sender: TelegramSender,
-    session: AsyncSession,
-    bot_uptime_seconds: float,
-    is_chat_run_active: Callable[[int], bool],
-) -> str:
-    """Render the gateway + per-conversation status reply for ``/status``.
-
-    Args:
-        sender: Normalized sender identity (carries ``chat_id`` for the
-            run-status lookup and ``thread_id`` for topic chats).
-        session: Async database session.
-        bot_uptime_seconds: Seconds since this worker booted (passed in so
-            the handler stays a pure function over its inputs).
-        is_chat_run_active: Predicate that returns whether ``chat_id`` has
-            an in-flight agent run on this worker. Process-local — a run
-            on another worker reads as idle here.
-
-    Returns:
-        Reply string the bot should send immediately.
-    """
-    nexus_user_id = await get_user_id_for_external(
-        provider=PROVIDER,
-        external_user_id=str(sender.user_id),
-        session=session,
-    )
-    if nexus_user_id is None:
-        return _STATUS_NOT_BOUND_MESSAGE
-
-    conversation = await get_or_create_telegram_conversation_full(
-        user_id=nexus_user_id,
-        session=session,
-        thread_id=sender.thread_id,
-    )
-    status = await get_conversation_status(conversation_id=conversation.id, session=session)
-    if status is None:
-        # The row was deleted between resolution and status read — extremely
-        # unlikely but render the gateway-only view rather than crashing.
-        return _STATUS_NO_CONVERSATION_MESSAGE.format(uptime=_format_duration(bot_uptime_seconds))
-
-    return _render_status_message(
-        bot_uptime_seconds=bot_uptime_seconds,
-        status=status,
-        run_active=is_chat_run_active(sender.chat_id),
-        thread_id=sender.thread_id,
-        now=_now_utc(),
-    )
+# ``handle_status_command`` + ``_render_status_message`` are imported
+# at the top of this module from ``app.integrations.telegram.status``
+# and re-exported, so historical call sites keep working.
